example/requirements.txt
Django
djangorestframework
drf-nested-routers
django-rest-offlinesync
